[fix] MyPage ViewController 파일 Extension, Layout 분류 및 폴더 재분류
