+++ B-umiOS/B-umiOS/MyPage/ViewController/MyRewardViewController/MyRewardViewController+Layout.swift
@@ -1,0 +1,27 @@
+//
+//  MyRewardViewController+Layout.swift
+//  B-umiOS
+//
+//  Created by 홍진석 on 2021/08/05.
+//
+
+import UIKit
+
+extension MyRewardViewController {
+    func setConstraints(){
+        view.addSubviews([myRewardCollectionView, errorView, errorLabel])
+        myRewardCollectionView.snp.makeConstraints { make in
+            make.top.equalToSuperview()
+            make.bottom.leading.trailing.equalToSuperview()
+        }
+        errorView.snp.makeConstraints { make in
+            make.top.equalToSuperview().inset(243 * SizeConstants.screenRatio)
+            make.centerX.equalToSuperview()
+        }
+        
+        errorLabel.snp.makeConstraints { make in
+            make.top.equalTo(errorView.snp.bottom).inset(10)
+            make.centerX.equalToSuperview()
+        }
+    }
+}
