+++ B-umiOS/B-umiOS/MyPage/ViewController/MyTrashBinViewController/MyTrashBinViewController+Extension.swift
@@ -1,0 +1,46 @@
+//
+//  MyTrashBinViewController+Extension.swift
+//  B-umiOS
+//
+//  Created by 홍진석 on 2021/08/05.
+//
+
+import UIKit
+
+// MARK: - UITableViewDelegate
+
+extension MyTrashBinViewController: UITableViewDelegate {
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        headerView
+    }
+    
+    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
+        72 * SizeConstants.screenRatio
+    }
+}
+
+// MARK: - UITableViewDataSource
+
+extension MyTrashBinViewController: UITableViewDataSource {
+    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+        return myTrashCan.count
+    }
+    
+    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        
+        guard let cell = tableView.dequeueReusableCell(withIdentifier: MyTrashBinTableViewCell.identifier, for: indexPath) as? MyTrashBinTableViewCell else {
+            return UITableViewCell()
+        }
+        cell.setTrashBinData(data: myTrashCan, index: indexPath.row)
+        return cell
+    }
+}
+
+// MARK: - Extension
+
+extension MyTrashBinViewController: viewDelegate {
+    func backgroundRemove() {
+        backgroundView.removeFromSuperview()
+    }
+}
+
