+++ B-umiOS/B-umiOS/MyPage/ViewController/MyTrashBinViewController/MyTrashBinViewController+Layout.swift
@@ -1,0 +1,41 @@
+//
+//  MyTrashBinViewController+Layout.swift
+//  B-umiOS
+//
+//  Created by 홍진석 on 2021/08/05.
+//
+
+extension MyTrashBinViewController {
+    func setConstraint(){
+        view.addSubviews([detailTableView, errorView, errorLabel, gradientView])
+        headerView.addSubviews([headerGardientBackground, settingButton])
+        
+        errorView.snp.makeConstraints { make in
+            make.top.equalToSuperview().inset(243 * SizeConstants.screenRatio)
+            make.centerX.equalToSuperview()
+        }
+        
+        errorLabel.snp.makeConstraints { make in
+            make.top.equalTo(errorView.snp.bottom).inset(10)
+            make.centerX.equalToSuperview()
+        }
+
+        headerGardientBackground.snp.makeConstraints { make in
+            make.top.leading.trailing.bottom.equalToSuperview()
+        }
+        settingButton.snp.makeConstraints { make in
+            make.trailing.equalToSuperview().inset(16 * SizeConstants.screenRatio)
+            make.centerY.equalToSuperview()
+        }
+
+        detailTableView.snp.makeConstraints { make in
+            make.top.leading.trailing.equalToSuperview()
+            make.bottom.equalTo(self.view.safeAreaLayoutGuide.snp.bottom)
+        }
+        
+        gradientView.snp.makeConstraints { make in
+            make.bottom.equalToSuperview()
+            make.leading.trailing.equalToSuperview()
+        }
+    }
+}
