+++ B-umiOS/B-umiOS/MyPage/ViewController/MyWritingViewController/MyWritingViewController+Layout.swift
@@ -1,0 +1,26 @@
+//
+//  MyWritingViewController+Layout.swift
+//  B-umiOS
+//
+//  Created by 홍진석 on 2021/08/05.
+//
+
+extension MyWritingViewController {
+    func setConstraints() {
+        view.addSubviews([myWritingCollectionView, errorView, errorLabel])
+        myWritingCollectionView.snp.makeConstraints { make in
+            make.top.equalToSuperview()
+            make.bottom.leading.trailing.equalToSuperview()
+        }
+        
+        errorView.snp.makeConstraints { make in
+            make.top.equalToSuperview().inset(243 * SizeConstants.screenRatio)
+            make.centerX.equalToSuperview()
+        }
+        
+        errorLabel.snp.makeConstraints { make in
+            make.top.equalTo(errorView.snp.bottom).inset(10)
+            make.centerX.equalToSuperview()
+        }
+    }
+}
